+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 ### 功能移除
 
+- 删除旧的使能缓存相关服务。
+  - com.dwarfeng.judge.stack.service.EnabledDriverInfoLookupService.java。
+  - com.dwarfeng.judge.stack.service.EnabledJudgerInfoLookupService.java。
+
 - 删除旧的数据评价相关机制。
   - 驱动机制。
   - 判断机制。